Clean up code
Change-type: patch
